backend/requirements.txt
django 
djangorestframework 
psycopg2-binary 
djangorestframework-simplejwt 
django-cors-headers
pytest-django
python-dotenv
pytest-html
drf-yasg